Rename testing auth manager
