+++ jest.config.js
@@ -3,6 +3,7 @@ module.exports = {
   testEnvironment: 'node',
   reporters: ['default', 'jest-junit'],
   testResultsProcessor: 'jest-junit',
+  testTimeout: 120000,
   globals: {
     'ts-jest': {
       tsConfig: './modules/tsconfig.settings.json'
