+++ modules/ipfs-cpinner-client/src/index.ts
@@ -1,4 +1,4 @@
-import AuthManager from './auth-manager'
+import AuthManager from './auth-manager/testing'
 import { AUTHENTICATION_ERROR, MAX_STORAGE_REACHED, SERVICE_MAX_STORAGE_REACHED, UNKNOWN_ERROR } from './constants'
 import {
   CreateContentPayload, CreateContentResponse,
