+++ modules/ipfs-cpinner-client/src/types.ts
@@ -1,4 +1,4 @@
-import AuthManager from './auth-manager'
+import AuthManager from './auth-manager/testing'
 import { IEncryptionManager } from './encryption-manager/types'
 export type GetContentPayload = { key: string }
 export type GetContentResponsePayload = { id: string, content: string }
