+++ modules/ipfs-cpinner-client/test/sandbox-env.test.ts
@@ -1,7 +1,7 @@
 import DataVaultWebClient from '../src'
 import { identityFactory, getEncryptionPublicKeyTestFn, decryptTestFn, customStorageFactory } from './util'
 import localStorageMockFactory from './localStorageMockFactory'
-import AuthManager from '../src/auth-manager'
+import AuthManager from '../src/auth-manager/testing'
 import EncryptionManager from '../src/encryption-manager/asymmetric'
 
 jest.setTimeout(12000)
